requirements.txt
Django==2.2.7
djangorestframework==3.10.3
psycopg2==2.8.4
psycopg2-pool==1.1
pytz==2019.3
sqlparse==0.3.0
